Implemented basic layout and added feedback to forms
The site now has a basic layout with a navbar and a footer. The navbar contains links to the login, register, user and logout page.
The landing page was scrapped and unified with the main page, this can be changed later if necessary.
If the main page is visited by a logged out user, the user page and logout links are hidden and instead login and register links are show (vice versa for logged in users)

The different forms now have some form of feedback in the form of a green/red box with a bit of text.