Added Bootstrap, JQuery and DataTables as local files, making the expected functionality actually work
The files are no longer loaded in via CDNs, which were giving me errors (altough they shouldn't).
To get a table to use the DataTables plugin, add a class 'display' to it.

If you want it to have mroe special features (autosort on column X, ...)
add a script after the table (just like the one in the head of base.html)
but instead of calling it on table.display call it on #tableID (don't forget to give a table id to the table)

Take a look at the examples/docs of DataTables, it has a lot of nice features: https://www.datatables.net/examples/index