Deleting datasets and tables now possible via web UI.
The order of the parameters in a function call to Dataloader was mixed up on my side.

Also changed request methods for delete buttons from DELETE to POST and send them to /path/delete instead. HTML forms just didn't want to play nicely.

We still don't check if the user deleting has rights to do so, but it shouldn't be a difficult addition. I'd prefer to merge these changes first before fixing new bugs (which may in turn create new ones)

Signed-off-by: Sergio Fenoll <sergio@fenoll.be>